Last commit. Fixing and implementation for UI design.

+++ app/src/main/java/com/assist/imobilandroidapp/screens/averageUser/screens/normalScreen/Classes/Listing.kt
@@ -14,7 +14,7 @@ data class Listing(
     @SerializedName("location")
     val location: ArrayList<String>? = null,
     @SerializedName("price")
-    val price: Int? = null,
+    val price: Double? = null,
     @SerializedName("status")
     val status: Int? = null,
     @SerializedName("images")
